Add devcontainer and GitHub Actions workflow
With this commit, we add a development container configuration and a GitHub Actions workflow to automate the building of Jupyter notebooks. The devcontainer setup allows for a consistent development environment, while the workflow ensures that notebooks are built and tested on every push to the repository.
This also includes updates to the project structure and some file cleanups.
Jupyter notebooks are now commited 'clean' and the output is not included in the repository.